pyproject.toml
[tool.poetry]
name = "test-case-simetra"
version = "0.1.0"
description = ""
authors = ["Dmitry Patin <multik.patin@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.6.3"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.28"
geoalchemy2 = "^0.14.6"
fastapi = {version = "^0.110.0", optional = true}
uvicorn = {version = "^0.27.1", extras = ["standard"], optional = true}
psycopg2-binary = {version = "^2.9.9", optional = true}
openpyxl = {version = "^3.1.2", optional = true}
alembic = {version = "^1.13.1", optional = true}

[tool.poetry.extras]
migrations = ["alembic"]
api = ["uvicorn", "fastapi"]
extractor = ["openpyxl", "psycopg2-binary"]

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.0"
pre-commit = "^3.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "UP", # pyupgrade
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "SIM",# flake8-simplify
]
ignore = [
    "B905",
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

fixable = ["ALL"]
unfixable = []

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "tests",
    "dist",
    "node_modules",
    "venv",
    "*/settings.py",
    "*/migrations/",
    "*/config/components/"
]
per-file-ignores = {}

line-length = 80

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

target-version = "py311"

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 60
